+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 - Fix missing Drupal version detection
 - Tokaido will now prompt for a Drupal version and path if it can't detect one automatically
 - Add the /private and sites/*/files folders to gitignore automatically
+- Add dependency checks for Linux